requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
openai
python-dotenv
chromadb